requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
dj-database-url==3.0.1
Django==4.2.24
django-cors-headers==4.8.0
django-jazzmin==3.0.1
django-unfold==0.65.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
dynamic-rest==2.3.0
gunicorn==23.0.0
hashids==1.3.1
idna==3.10
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.10.0
python-dotenv
Pillow
dj-database-url
django-import-export
openpyxl


